+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this changelog file.
 
+## [Unreleased]
+### Added
+- Support for remote Frida server
+
 ## [1.0] - 2023-12-01
 ### Added
 - WinSock, OpenSSL and SChannel libraries